feat(videocore): characters
- fix qbit type mismatch

+++ internal/torrent_clients/qbittorrent/model/torrent.go
@@ -56,7 +56,7 @@ type Torrent struct {
 	InfohashV2        string  `json:"infohash_v2"`
 	LastActivity      int     `json:"last_activity"`
 	MagnetUri         string  `json:"magnet_uri"`
-	MaxRatio          int     `json:"max_ratio"`
+	MaxRatio          float64 `json:"max_ratio"`
 	MaxSeedingTime    int     `json:"max_seeding_time"`
 	RatioLimit        int     `json:"ratio_limit"`
 	SavePath          string  `json:"save_path"`
